feat(132): standardize on a format for standard methods
The standard method AEPs have a bit of inconsistency
in the format and the examples they provide.

This PR proposes a format to normalize them all to, including
the following headings:

- Overview: a brief explanation of the method
- Guidance: a brief explanation of the purpose
  - Operation: an example operation, with must/should guidance
  - Request: an example request, with must/should guidance
  - Response: an example request, with must/should guidance
  - Errors: explanation on errors, if applicable
  - (additional guidance sections specific to the AEP)
- Interface Definitions: a full example of the operation, request, and response.

This will result in some incomplete sections: those will be filled out via
follow-up PRs to keep this from becoming too massive.