feat(134): replace etag guidance with 154 reference
Previously, there was bespoke guidance in the Update AEP that
refered to using a custom ETag field in the resource for
validating previous resource state.

This guidance somewhat duplicates the guidance in RFC 9110 
and AEP-154 which states that it should be adhered to. This 
duplication can lead to partial support of both, which in turn
will make client integration more difficult.

Therefore, removing this guidance. Fixes #241