feat(164): specify create overwrites soft deleted
To help ensure that declarative clients do not have a
degraded user experience with resources that implement 
soft-delete, specify that create overwrites an existing
soft-deleted resource.

This ensures that a declarative tool can still properly
apply a resource, overwriting it instead of requiring an
out of band force delete or having the user modify the
ID of the resource.

fixes #111 (which includes some significant debate and
context on why this approach was chosen).