fixing out of date guidance (#392)
* fixing out of date guidance

Updated contributing guidelines to clarify the process for editing AEP rules and submitting PRs. Removes out-of-date information about the site generator and points to our dedicated repo.

* adding info about our comms channels

slack and gcal moved from readme

* Pointing to the aep.dev site content

Rather than add clicks to journeys (by visiting a .md file and then the website), this helps people get where they need to be faster.

* re-add serve.sh

adds back instructions for the script for running locally and fixes the broken `fix.py` link

* addresses markdown linting issues

Because I was working on a Chromebook, I didn't have an easy way to run the linter on markdown, and the GitHub workflow doesn't give enough insight into what violations it encountered. 

Luckily there is an online version of [Prettier](https://prettier.io/playground), so I've used that.

* fix linting issue